Changed the feature so that the default behavior is to delay a single photo for as long as possible instead of posting it immediately.  (I think this will simply the math in a number of places down the road.)
I introduced a constant MAXIMUM_UPLOAD_DELAY in ScheduledPhotosHelper which is currently set to 7 days.  I'm not sure that's the best place for the constant, but it works for now.  I then needed to include ScheduledPhotosHelper in a both ScheduledPhoto and in the steps file for ScheduledPhoto so I could use the new constant.

I also wrote a new scenario where I'm scheduling my first three photos.