Got the calculations for setting and updating the upload time for scheduled photos worked out.  It took a fair bit longer than I expected because I ended up struggling with where to put the logic to update the upload times for previously queued photos.  In the end it turned out that I wanted the update action in after_create, because that meant that I could assume the new ScheduledPhoto object was created and saved to the database before I start messing around.
I had to make several changes to the feature file and steps for two reasons.  First, there were scenarios that referenced absolute offsets instead of doing things in terms of the maximum upload delay.  Second, I had started out assuming that the first photo would be posted immediately if there was no prior history; I've since decided that was at best an awkward idea, and that it made more sense to have the first photo uploaded in MAXIMUM_UPLOAD_DELAY / (# of photos in queue).